+++ tests/AnalysisTest.php
@@ -30,7 +30,7 @@ class AnalysisTest extends TestCase
      *
      * @return string[]
      */
-    protected function getPaths()
+    protected static function getPaths()
     {
         return [
             realpath(__DIR__.'/../src'),
@@ -43,7 +43,7 @@ class AnalysisTest extends TestCase
      *
      * @return string[]
      */
-    protected function getIgnored()
+    protected static function getIgnored()
     {
         return ['Foo\\Bar', 'Foo\\Baz'];
     }